+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Introduce flaw ownership through task management system (OSIDB-69)
+- Implement task rejection in Taskman (OSIDB-74)
 
 ### Changed
 - Integrate Jira tracker collector with collector framework (OSIDB-576)